MITAB: fix various warnings raised by MSVC (https://trac.osgeo.org/gdal/ticket/5652)
git-svn-id: https://svn.osgeo.org/gdal/trunk@27695 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/mitab/mitab_idfile.cpp
@@ -208,7 +208,7 @@ int TABIDFile::Open(const char *pszFname, TABAccess eAccess)
             return -1;
         }
 
-        m_nMaxId = sStatBuf.st_size/4;
+        m_nMaxId = (int)(sStatBuf.st_size/4);
         m_nBlockSize = MIN(1024, m_nMaxId*4);
 
         /*-------------------------------------------------------------
